Upload TL1 and TL2
Note: these are messy and will be updated in the near future.

TL1: The first portion of the translation layer. This file reads PowerGenome settings files and populates csv files in a very similar manner to how it would for GenX. Some Temoa-specific modifications are made.
TL2: This portion of the translation layer translates the csv files into a SQLite file that is used as the input file for Temoa runs.